Rename map components and map folder
