+++ src/views/MapView/utils/drawIcon.js
