+++ src/views/MapView/utils/swapCoordinates.js
