+++ src/views/ServiceView/ServiceView.js
@@ -8,8 +8,8 @@ import SearchBar from '../../components/SearchBar';
 import TitleBar from '../../components/TitleBar/TitleBar';
 import { DesktopComponent, MobileComponent } from '../../layouts/WrapperComponents/WrapperComponents';
 import { generatePath } from '../../utils/path';
-import { drawServiceIcon } from '../Map/utils/drawIcon';
-import { fitUnitsToMap } from '../Map/utils/mapActions';
+import { drawServiceIcon } from '../MapView/utils/drawIcon';
+import { fitUnitsToMap } from '../MapView/utils/mapActions';
 import ResultList from '../../components/Lists/ResultList';
 import Loading from '../../components/Loading/Loading';
 import Container from '../../components/Container/Container';
